Actualización v0.3

+++ app/routes.py
@@ -1,6 +1,6 @@
 from flask import render_template, flash, redirect, session, request
 from app import app
-from app.forms import LoginForm, RegisterForm
+from app.forms import LoginForm, RegisterForm, ModifyForm, AgregarLugar
 import pymysql.cursors
 
 # Connect to the database
@@ -39,7 +39,7 @@ def bienvenido():
                     session['nombre'] = userdata['nombre']
                     session['tipo'] = userdata['tipo']
                     return redirect('/principal')
-    return render_template('bienvenido.html', form=form)
+    return render_template('bienvenido.html', title='¡Bienvenido a UruGuia!', form=form)
 
 @app.route('/index/registro', methods=['GET', 'POST'])
 def registro():
@@ -94,26 +94,161 @@ def registro():
                 return redirect('/index') 
             except NameError:    
                     print('Algo malo acaba de ocurrir: ' + NameError)         
-    return render_template('registro.html', form=form,tipo=nuevo['tipo'])
+    return render_template('registro.html', title="Registrar un nuevo usuario", form=form, tipo=nuevo['tipo'])
 
 @app.route('/principal')
 def index():
     if session.get('logueado'):
-        print('ok')
+        if session['tipo']=='turista':
+            pass
+        else:
+            try:
+                with connection.cursor() as cursor:
+                    #CONSIGUE lugares DE LA EMPRESA
+                    query = """
+                            SELECT l.nombre,
+                                   l.descripcion,
+                                   l.ubicacion,
+                                   l.tipo,
+                                   l.horario,
+                                   l.fecha
+                            FROM tiene t
+                            INNER JOIN lugar l
+                            ON t.ide=l.ide
+                            WHERE id=%s
+                            """
+                    cursor.execute(query,(session['id_usuario']))
+                    lugares = cursor.fetchall()
+                connection.commit()
+            except:
+                # Debe devolver un objeto vacío para listar
+                return redirect('/')      
     else:
         session.clear()
         return redirect('/')
-    return render_template('principal.html')
+    return render_template('principal.html', title="Página principal", lugares=lugares)
 
 @app.route('/principal/usuario')
 def usuario():
-    if session.get('logueado'):
-        print('ok')
-    else:
+    if not session.get('logueado'):
         session.clear()
         return redirect('/')
-    return render_template('usuario.html')
+    return render_template('usuario.html', title='Datos de la cuenta')
 
+@app.route('/principal/usuario/modificar', methods=['GET','POST'])
+def modificarUsuario():
+    if not session.get('logueado'):
+        session.clear()
+        return redirect('/')
+    else:
+        form=ModifyForm()
+        try:
+            if form.validate_on_submit():
+                nombre = form.nombre.data
+                password = form.password.data
+                edad = form.edad.data
+                pais = form.pais.data
+                nombreEmpresa = form.nombreEmpresa.data
+                with connection.cursor() as cursor:
+                    query="""
+                        UPDATE usuario
+                        SET nombre=%s,
+                            contrasena=%s
+                        WHERE id_usuario=""" + str(session['id_usuario'])
+                    cursor.execute(query,(nombre,password))
+                    if session['tipo']=='turista':
+                        query="""
+                            UPDATE turista
+                            SET edad=%s,
+                                pais_origen=%s
+                            WHERE id=""" + str(session['id_usuario'])
+                        cursor.execute(query,(edad,pais))
+                    else:
+                        query="""
+                            UPDATE empresa
+                            SET nombre=%s
+                            WHERE id=""" + str(session['id_usuario'])
+                        cursor.execute(query,(nombreEmpresa))
+                connection.commit()
+                session['nombre'] = nombre
+                return redirect("/principal/usuario")
+            with connection.cursor() as cursor:
+                query="SELECT nombre, contrasena FROM usuario WHERE id_usuario=" + str(session['id_usuario'])
+                cursor.execute(query)
+                usuario=cursor.fetchone()
+                if session['tipo']=='turista':
+                    query="SELECT edad, pais_origen FROM turista WHERE id=" + str(session['id_usuario'])
+                    cursor.execute(query)
+                    turista=cursor.fetchone()
+                    empresa={"nombre":"Ninguno"}
+                else:
+                    query="SELECT nombre FROM empresa WHERE id=" + str(session['id_usuario'])
+                    cursor.execute(query)
+                    empresa=cursor.fetchone()
+                    turista={"edad":"0", "pais_origen":"Ninguno"}
+            connection.commit()
+        except:
+            return redirect('/')                    
+    print(session["tipo"])
+    return render_template('modificar_usuario.html', title="Modificar datos de la cuenta", form=form, contrasena=usuario['contrasena'], edad=turista['edad'], pais=turista['pais_origen'], nombreEmpresa=empresa['nombre']) 
+
+@app.route('/principal/agregar_lugar', methods=['GET','POST'])
+def agregar_lugar():
+    if not session.get('logueado'):
+        session.clear()
+        return redirect('/')
+    else:
+        form=AgregarLugar()
+        try:
+            if form.validate_on_submit():
+                nombre = form.nombre.data
+                descripcion = form.descripcion.data
+                ubicacion = form.ubicacion.data
+                categoria = form.categoria.data
+                tipo = form.tipo.data
+                horario = form.horario.data
+                fecha = form.fecha.data
+                with connection.cursor() as cursor:
+                    #INSERTA EN lugar
+                    query = """
+                            INSERT INTO lugar
+                            (nombre, descripcion, ubicacion, tipo, horario, fecha)
+                            VALUES
+                            (%s,%s,%s,%s,%s, %s)
+                            """
+                    cursor.execute(query,(nombre, descripcion, ubicacion, tipo, horario, fecha))
+                    
+                    #CONSIGUE ide DE lugar
+                    query = """
+                            SELECT ide FROM lugar
+                            WHERE nombre=%s
+                            """
+                    cursor.execute(query,(nombre))
+                    lugar = cursor.fetchone()
+
+                    #INSERTA EN pertenece_a BASANDOSE EN EL ide
+                    query = """
+                            INSERT INTO pertenece_a
+                            (ide, idc)
+                            VALUES
+                            (%s,%s)
+                            """
+                    cursor.execute(query,(lugar['ide'],categoria))
+
+                    #INSERTA EN tiene BASANDOSE EN ide Y id
+                    query = """
+                            INSERT INTO tiene
+                            (ide, id)
+                            VALUES
+                            (%s,%s)
+                            """
+                    cursor.execute(query,(lugar['ide'],session['id_usuario']))
+
+                connection.commit()
+                return redirect("/principal")
+        except:
+            return redirect('/')      
+    return render_template('agregar_lugar.html', title="Registrar un establecimiento o evento", form=form)
 
 @app.route('/logout')
 def logout():
@@ -121,22 +256,23 @@ def logout():
         idu = request.args.to_dict()
         iduINT = int(idu['id_usuario'])
         ses = int(session['id_usuario'])
-        with connection.cursor() as cursor:
-            query = """
-                    DELETE FROM usuario WHERE id_usuario=%s
-                    """ 
-            cursor.execute(query,(ses))
-            if session['tipo']=='turista':
+        if iduINT==ses:
+            with connection.cursor() as cursor:
                 query = """
-                        DELETE FROM turista WHERE id=%s
+                        DELETE FROM usuario WHERE id_usuario=%s
                         """ 
                 cursor.execute(query,(ses))
-            else:
-                query = """
-                    DELETE FROM empresa WHERE id=%s
-                    """ 
-                cursor.execute(query,(ses))
-        connection.commit()
+                if session['tipo']=='turista':
+                    query = """
+                            DELETE FROM turista WHERE id=%s
+                            """ 
+                    cursor.execute(query,(ses))
+                else:
+                    query = """
+                        DELETE FROM empresa WHERE id=%s
+                        """ 
+                    cursor.execute(query,(ses))
+            connection.commit()
     except:
         print("Logout")           
     session.clear()
